refactor: remove controller layer
