+++ server/app/routes/tool_routes.py
@@ -3,7 +3,8 @@ from fastapi import APIRouter, Depends, HTTPException, status, Query
 from typing import List, Optional, Union
 from uuid import UUID
 
-from ..controllers.tool_controller import tool_controller # Use the instance
+from ..crud import tool_crud
+from ..supabase_client import create_supabase_client_with_jwt
 from ..schemas.tool_schemas import (
     ToolCreateRequestSchema,
     CustomApiToolCreateRequestSchema,
@@ -11,7 +12,7 @@ from ..schemas.tool_schemas import (
     ToolUpdateRequestSchema,
     PaginatedToolResponseSchema
 )
-from ..security import get_current_user_and_token, CurrentUser # Updated dependency
+from ..security import get_current_user_and_token, CurrentUser
 
 router = APIRouter(
     prefix="/tools",
@@ -22,25 +23,16 @@ router = APIRouter(
 @router.post("/", response_model=ToolResponseSchema, status_code=status.HTTP_201_CREATED)
 async def create_new_tool(
     tool_payload: Union[ToolCreateRequestSchema, CustomApiToolCreateRequestSchema],
-    current_user: CurrentUser = Depends(get_current_user_and_token) # Updated dependency
+    current_user: CurrentUser = Depends(get_current_user_and_token)
 ):
-    """
-    Create a new tool with its parameters.
-    The user_id and jwt_token are extracted from the JWT token.
-    FastAPI will automatically handle the Union type based on the payload.
-    """
-    try:
-        return await tool_controller.create_tool(
-            tool_data=tool_payload, 
-            user_id=current_user.id, 
-            jwt_token=current_user.token
-        )
-    except ValueError:
-        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="Invalid user ID format.")
-    except HTTPException as e:
-        raise e
-    except Exception as e:
-        raise HTTPException(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR, detail=f"An unexpected error occurred: {str(e)}")
+    """Create a new tool."""
+    db = create_supabase_client_with_jwt(current_user.token)
+    tool_dict = tool_payload.model_dump(exclude_unset=True)
+    tool_dict["owner_id"] = str(current_user.id)
+    response = await db.table("tools").insert(tool_dict).execute()
+    if not response.data:
+        raise HTTPException(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR, detail="Failed to create tool")
+    return ToolResponseSchema(**response.data[0])
 
 
 @router.get("/", response_model=PaginatedToolResponseSchema)
@@ -54,91 +46,65 @@ async def list_all_tools(
     List all tools available to the user, including system tools.
     Supports pagination.
     """
-    try:
-        return await tool_controller.list_tools(
-            user_id=current_user.id,
-            jwt_token=current_user.token,
-            page=page,
-            size=size,
-            include_system_tools=include_system_tools
-        )
-    except ValueError:
-        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="Invalid user ID format.")
-    except HTTPException as e:
-        raise e
-    except Exception as e:
-        raise HTTPException(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR, detail=f"An unexpected error occurred: {str(e)}")
+    db = create_supabase_client_with_jwt(current_user.token)
+    offset = (page - 1) * size
+    response = await db.table("tools").select("*", count="exact").range(offset, offset + size - 1).execute()
+    tools = [ToolResponseSchema(**t) for t in response.data] if response.data else []
+    total_count = response.count or len(tools)
+    return PaginatedToolResponseSchema(tools=tools, total_count=total_count, skip=offset, limit=size)
 
 
 @router.get("/{tool_id}", response_model=ToolResponseSchema)
 async def get_single_tool(
-    tool_id: UUID, 
-    current_user: CurrentUser = Depends(get_current_user_and_token) # Updated dependency
+    tool_id: UUID,
+    current_user: CurrentUser = Depends(get_current_user_and_token)
 ):
     """
     Retrieve a single tool by its ID.
     Ensures the user has permission to view the tool (own or system tool).
     """
-    try:
-        tool = await tool_controller.get_tool(
-            tool_id=tool_id, 
-            user_id=current_user.id, 
-            jwt_token=current_user.token
-        )
-        return tool
-    except ValueError:
-        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="Invalid ID format.")
-    except HTTPException as e:
-        raise e
-    except Exception as e:
-        raise HTTPException(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR, detail=f"An unexpected error occurred: {str(e)}")
+    db = create_supabase_client_with_jwt(current_user.token)
+    response = await db.table("tools").select("*").eq("id", str(tool_id)).maybe_single().execute()
+    if not response.data:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Tool not found")
+    return ToolResponseSchema(**response.data)
 
 
 @router.put("/{tool_id}", response_model=ToolResponseSchema)
 async def update_existing_tool(
     tool_id: UUID,
     tool_update_payload: ToolUpdateRequestSchema,
-    current_user: CurrentUser = Depends(get_current_user_and_token) # Updated dependency
+    current_user: CurrentUser = Depends(get_current_user_and_token)
 ):
     """
     Update an existing tool.
     Ensures the tool belongs to the user and is not a system tool.
     """
-    try:
-        updated_tool = await tool_controller.update_tool(
-            tool_id=tool_id,
-            tool_data=tool_update_payload,
-            user_id=current_user.id,
-            jwt_token=current_user.token
-        )
-        return updated_tool
-    except ValueError:
-        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="Invalid ID format for tool or user.")
-    except HTTPException as e:
-        raise e
-    except Exception as e:
-        raise HTTPException(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR, detail=f"An unexpected error occurred: {str(e)}")
+    db = create_supabase_client_with_jwt(current_user.token)
+    update_data = tool_update_payload.model_dump(exclude_unset=True)
+    if not update_data:
+        response = await db.table("tools").select("*").eq("id", str(tool_id)).maybe_single().execute()
+        if not response.data:
+            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Tool not found")
+        return ToolResponseSchema(**response.data)
+
+    response = await db.table("tools").update(update_data).eq("id", str(tool_id)).execute()
+    if not response.data:
+        raise HTTPException(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR, detail="Failed to update tool")
+    return ToolResponseSchema(**response.data[0])
 
 
 @router.delete("/{tool_id}", status_code=status.HTTP_204_NO_CONTENT)
 async def delete_existing_tool(
     tool_id: UUID,
-    current_user: CurrentUser = Depends(get_current_user_and_token) # Updated dependency
+    current_user: CurrentUser = Depends(get_current_user_and_token)
 ):
     """
     Delete an existing tool.
     Ensures the tool belongs to the user and is not a system tool.
     """
-    try:
-        await tool_controller.delete_tool(
-            tool_id=tool_id, 
-            user_id=current_user.id, 
-            jwt_token=current_user.token
-        )
-        # No return needed for 204
-    except ValueError:
-        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="Invalid ID format for tool or user.")
-    except HTTPException as e:
-        raise e
-    except Exception as e:
-        raise HTTPException(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR, detail=f"An unexpected error occurred: {str(e)}")
+    db = create_supabase_client_with_jwt(current_user.token)
+    response = await db.table("tools").delete().eq("id", str(tool_id)).execute()
+    if not response.data:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Tool not found")
+    return
